+++ java/samples/sample-code/src/main/java/com/microsoft/semantickernel/samples/syntaxexamples/Example69_MutableKernelPlugin.java
@@ -8,6 +8,7 @@ import com.microsoft.semantickernel.plugin.annotations.DefineKernelFunction;
 public class Example69_MutableKernelPlugin {
 
     public static class Time {
+
         @DefineKernelFunction(name = "date")
         public String date() {
             return "2021-09-01";
@@ -17,21 +18,22 @@ public class Example69_MutableKernelPlugin {
     /**
      * Mutable KernelPlugin example
      * <p>
-     * KernelFunction can be added directly to the function collection of the KernelPlugin
-     * by using KernelPlugin.getFunctions and putting the corresponding KernelFunction
+     * KernelFunction can be added directly to the function collection of the KernelPlugin by using
+     * KernelPlugin.getFunctions and putting the corresponding KernelFunction
      */
     public static void main(String[] args) throws NoSuchMethodException {
         System.out.println("======== Example69_MutableKernelPlugin ========");
 
         KernelPlugin plugin = new KernelPlugin("Plugin", "Mutable plugin", null);
         plugin.getFunctions().put("dateFunction", KernelFunctionFactory.createFromMethod(
-                Time.class.getMethod("date"), new Time(),"date", null, null, null));
+            Time.class.getMethod("date"), new Time(), "date", null, null, null));
 
         Kernel kernel = Kernel.builder().build();
         kernel.getPlugins().add(plugin);
 
-        var result = kernel.invokeAsync(kernel.getPlugins().getFunction("Plugin", "dateFunction"), null, String.class).block();
+        var result = kernel.invokeAsync(kernel.getPlugins().getFunction("Plugin", "dateFunction"),
+            null, String.class).block();
 
-        System.out.println("Result: " + result.getValue());
+        System.out.println("Result: " + result.getResult());
     }
 }
